+++ src/theme/ITheme.ts
@@ -46,13 +46,18 @@ export interface IThemeColors {
     base: IThemeBaseColors;
 
     alethioLogo: string;
-    toolbarIcon: string;
-    toolbarIconHover: string;
+    toolbar: {
+        icon: {
+            normal: string;
+            hover: string;
+            active: string;
+            activeHover: string;
+        }
+    };
     toolbarAlethioIcon: string;
     toolbarAlethioIconHover: string;
 
     sidebarBg: string;
-    sidebarBorder: string;
     sidebarPageTitle: string;
     mainContentBg: string;